requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
pandas>=2.0.3
transformers>=4.34.0
torch>=2.0.1
numpy>=1.24.3
googlesearch-python>=1.2.3